Renamed symbtrdataextractor/structure_graph.py -> symbtrdataextractor/GraphOperations.py
